Move constants to constants.py and document them
